Move add_partition_keys_header to headers module
