+++ sdk/cosmos/src/requests/create_reference_attachment_builder.rs
@@ -205,7 +205,7 @@ impl<'a, 'b> CreateReferenceAttachmentBuilder<'a, 'b, Yes, Yes> {
         req = ActivityIdOption::add_header(self, req);
         req = ConsistencyLevelOption::add_header(self, req);
 
-        req = crate::add_partition_keys_header(
+        req = crate::headers::add_partition_keys_header(
             self.attachment_client.document_client().partition_keys(),
             req,
         );
