+++ sdk/cosmos/src/requests/delete_attachment_builder.rs
@@ -111,7 +111,7 @@ impl<'a, 'b> DeleteAttachmentBuilder<'a, 'b> {
         req = ActivityIdOption::add_header(self, req);
         req = ConsistencyLevelOption::add_header(self, req);
 
-        req = crate::add_partition_keys_header(
+        req = crate::headers::add_partition_keys_header(
             self.attachment_client.document_client().partition_keys(),
             req,
         );
