+++ sdk/cosmos/src/requests/delete_document_builder.rs
@@ -157,7 +157,7 @@ impl<'a> DeleteDocumentBuilder<'a> {
         req = ConsistencyLevelOption::add_header(self, req);
         req = AllowTentativeWritesOption::add_header(self, req);
 
-        req = crate::add_partition_keys_header(self.document_client.partition_keys(), req);
+        req = crate::headers::add_partition_keys_header(self.document_client.partition_keys(), req);
 
         let req = req.body(EMPTY_BODY.as_ref())?;
         debug!("{:?}", req);
