+++ sdk/cosmos/src/requests/get_document_builder.rs
@@ -139,7 +139,7 @@ impl<'a, 'b> GetDocumentBuilder<'a, 'b> {
         req = ActivityIdOption::add_header(self, req);
         req = ConsistencyLevelOption::add_header(self, req);
 
-        req = crate::add_partition_keys_header(self.document_client.partition_keys(), req);
+        req = crate::headers::add_partition_keys_header(self.document_client.partition_keys(), req);
 
         let req = req.body(EMPTY_BODY.as_ref())?;
