+++ sdk/cosmos/src/requests/replace_slug_attachment_builder.rs
@@ -231,7 +231,7 @@ impl<'a, 'b> ReplaceSlugAttachmentBuilder<'a, 'b, Yes, Yes> {
         req = ActivityIdOption::add_header(self, req);
         req = ConsistencyLevelOption::add_header(self, req);
 
-        req = crate::add_partition_keys_header(
+        req = crate::headers::add_partition_keys_header(
             self.attachment_client.document_client().partition_keys(),
             req,
         );
